requirements.txt
bcrypt==5.0.0
charset-normalizer==3.4.4
mysql-connector-python==9.4.0
mysqlclient==2.2.7
prettytable==3.16.0
pwinput==1.0.3
python-decouple==3.8
requests==2.32.5
rut-chile==2.0.1
SQLAlchemy==2.0.44
sqlparse==0.5.3